pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = ""
authors = [
    {name = "Fazley Rabbi",email = "sfazleyrabbi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10.0,<3.12"
dependencies = [
    "langchain-openai (>=0.3.8,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-anthropic (>=0.3.9,<0.4.0)",
    "langchain-google-genai (>=2.1.0,<3.0.0)",
    "langchain-google-firestore (>=0.5.0,<0.6.0)",
    "firestore (>=0.0.8,<0.0.9)",
    "chromadb (>=0.6.3,<0.7.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "firecrawl-py (>=1.13.5,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "tavily-python (>=0.5.1,<0.6.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
